# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [Auth]
      summary: Get the current account
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Authenticated user or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /login:
    get:
      operationId: beginBrowserLogin
      tags: [Auth]
      summary: Start browser login
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OIDC provider
...
  /messages/{messageId}/feedback:
    post:
      operationId: createMessageFeedback
      tags: [feedback]
      summary: Rate an assistant response
      parameters:
        - $ref: "#/components/parameters/MessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Saved feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageFeedback"
        "404":
          description: Message not found
...
  /knowledge/articles:
    get:
      operationId: listKnowledgeArticles
      tags: [knowledge]
      summary: Search knowledge articles
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, approved, archived]
      responses:
        "200":
          description: Knowledge articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeArticle"
    post:
      operationId: createKnowledgeArticle
      tags: [knowledge]
      summary: Create a knowledge article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeArticleInput"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeArticle"
  /knowledge/articles/{articleId}:
    patch:
      operationId: updateKnowledgeArticle
      tags: [knowledge]
      summary: Update a knowledge article
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeArticleUpdate"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeArticle"
        "404":
          description: Article not found
  /knowledge/articles/{articleId}/approve:
    post:
      operationId: approveKnowledgeArticle
      tags: [knowledge]
      summary: Approve a knowledge article
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200":
          description: Approved article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeArticle"
        "404":
          description: Article not found
  /knowledge/articles/{articleId}/archive:
    post:
      operationId: archiveKnowledgeArticle
      tags: [knowledge]
      summary: Archive a knowledge article
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200":
          description: Archived article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeArticle"
        "404":
          description: Article not found
components:
  parameters:
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie
      schema:
        type: string
...
    MessageId:
      name: messageId
      in: path
      required: true
      schema:
        type: string
    ArticleId:
      name: articleId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, role]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
          format: email
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]
        role:
          type: string
          enum: [customer, support, admin, system]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
...
    FeedbackInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: string
          enum: [helpful, not_helpful]
        feedback:
          type: ["string", "null"]
          maxLength: 1000
        score:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
    MessageFeedback:
      type: object
      required: [id, messageId, rating, feedback, createdAt]
      properties:
        id:
          type: string
        messageId:
          type: string
        rating:
          type: string
          enum: [helpful, not_helpful]
        score:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        feedback:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    KnowledgeArticle:
      type: object
      required: [id, title, category, content, version, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        content:
          type: string
        version:
          type: integer
        status:
          type: string
          enum: [draft, approved, archived]
        effectiveFrom:
          type: ["string", "null"]
          format: date-time
        effectiveUntil:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    KnowledgeArticleInput:
      type: object
      required: [title, category, content]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 180
        category:
          type: string
          minLength: 1
          maxLength: 60
        content:
          type: string
          minLength: 1
          maxLength: 20000
        status:
          type: string
          enum: [draft, approved]
    KnowledgeArticleUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 180
        category:
          type: string
          minLength: 1
          maxLength: 60
        content:
          type: string
          minLength: 1
          maxLength: 20000